feat(trip): AddStopModal → /trip/:id/add-stop?day=N page (modal-to-fullpage migration v2.19.7)
DESIGN.md 2026-05-03「Modal vs Full Page」規則第 4 個遷移 PR (PR series 4/4)：
AddStopModal → AddStopPage @ /trip/:tripId/add-stop?day=N。同 pattern 套用
PR #428/#429/#430，這是 4 個違規 modal 的最後一個。

What changed:
- NEW src/pages/AddStopPage.tsx (~1170 LOC) — AppShell + sticky TitleBar
  (返回 + 「加入景點」+ 「完成」 action with .tp-titlebar-action.is-primary)
  + content (3 tab: 搜尋/收藏/自訂 + region pill + filter sheet + POI grid +
  custom form) + bottom sticky bar (counter + 取消 + 完成)。Form 邏輯 / state /
  handlers 完全沿用原 AddStopModal v2.19.0 (含 6 region option + 5 category
  subtab + 3 saved sub-filter + custom form 4 fields)。
- DEL src/components/trip/AddStopModal.tsx (~1174 LOC)
- Route: /trip/:tripId/add-stop mounted under TripLayout，dayNum 從 ?day= URL
  searchParam 讀取 (URL deep-linkable)。
- TripPage caller 改 navigate(`/trip/:id/add-stop?day=N`) 取代 mount portal。
  順手 DEL deriveAddStopRegion helper (region selector 自己處理 default region)
  + DEL inline DAY label 計算 (deriveDayLabel 搬進 page，weekday 改用 API
  row.day_of_week 直取，取代 ['日','一',…] hardcode 陣列)。

何以全頁:
- 瀏覽器 back 鍵直接取消（不需 X close button + Esc handler + backdrop click）
- /trip/:id/add-stop?day=N URL deep-linkable（書籤、隊友分享）
- mobile 不被 modal 高度限制（3 tab 內容多 + region menu / filter sheet 在
  modal 內難放）
- 重新整理保留位置
- a11y 簡化（無需 focus trap、aria-modal、Escape handler）
- AddStopModal 是 4 個違規中最大 + 最複雜的 — 9 種 form control + 多 tab +
  動態 dropdown menu / sheet 都在 modal 裡，最該走全頁。

Tests:
- DEL tests/unit/add-stop-modal.test.tsx (modal-specific render / Esc /
  backdrop click，page 模式不適用，重寫成本 > 重新覆蓋價值，走 e2e 為主)
- RENAME tests/unit/add-stop-modal-region-filter.test.ts →
  add-stop-page-region-filter.test.ts，readFile path 改新 page，dayLabel
  rule 對齊新 deriveDayLabel + day_of_week 用法，加 dayNum useSearchParams 驗證
- RENAME tests/e2e/add-stop-modal.spec.js → add-stop-page.spec.js，trigger
  改驗 URL 跳轉 + page testid，關閉改驗 TitleBar 返回 + 加 TitleBar /
  bottom bar 同步 disabled state spec
- mockup-typography-compliance.test.ts: AddStopModal title 700 weight 規則
  改驗 page 用 TitleBar + tokens.css .tp-titlebar-title 700 weight (page
  不再有 inline h2)
- 1299 unit + 603 API 全綠，typecheck 乾淨

Migration progress (DESIGN.md 2026-05-03 audit table):
- ✅ EditTripModal → /trip/:id/edit (PR #428, v2.19.4)
- ✅ NewTripModal → /trips/new (PR #429, v2.19.5)
- ✅ EntryActionPopover → /trip/:id/stop/:eid/(copy|move) (PR #430, v2.19.6)
- ✅ AddStopModal → /trip/:id/add-stop?day=N (本次, v2.19.7) — 完成 4/4

Audit 剩餘 (低優先 / 不同類型):
- DeveloperAppsPage create-app modal (9+ field form, 預估走 /developer/apps/new)
- ExplorePage trip-picker modal (chooser 性質，評估 popover vs page)
- AccountPage logout / ConnectedAppsPage revoke / DeveloperAppsPage
  secret-display 改用標準化 ConfirmModal (P3 confirm-style cleanup)
