+++ VERSION
@@ -1,1 +1,1 @@
-2.19.6
+2.19.7
