+++ tests/e2e/add-stop-page.spec.js
@@ -1,10 +1,12 @@
 // @ts-check
 /**
- * AddStopModal E2E — Section 3 (terracotta-add-stop-modal)
+ * AddStopPage E2E — Section 3 (terracotta-add-stop-modal mockup)
  *
- * 驗 mockup section 14：trip-level「+ 加景點」trigger 開 modal，3 tab
- * (搜尋/收藏/自訂) + 5 subtab + footer counter + 自訂 form 提交 → timeline
- * 出現 entry。
+ * 2026-05-03 modal-to-fullpage migration: 原 AddStopModal 改 /trip/:id/add-stop?day=N
+ * 全頁。trigger 從 TitleBar「+ 加景點」按鈕走 navigate，不再 mount portal。
+ *
+ * 驗 mockup section 14：navigate 後 page render，3 tab (搜尋/收藏/自訂)
+ * + 5 subtab + 自訂 form + counter + back via 瀏覽器 history。
  */
 import { test, expect } from '@playwright/test';
 const { setupApiMocks } = require('./api-mocks');
@@ -13,18 +15,19 @@ test.beforeEach(async ({ page }) => {
   await setupApiMocks(page);
 });
 
-test.describe('AddStopModal — Section 3', () => {
-  test('TripPage TitleBar 「加景點」 button trigger 開 modal', async ({ page }) => {
+test.describe('AddStopPage — Section 3 (modal-to-fullpage migration)', () => {
+  test('TripPage TitleBar「加景點」 button → navigate 到 /add-stop?day=N page', async ({ page }) => {
     await page.goto('/trip/okinawa-trip-2026-Ray');
     const trigger = page.getByTestId('trip-add-stop-trigger');
     await expect(trigger).toBeVisible();
     await trigger.click();
-    await expect(page.getByTestId('add-stop-modal')).toBeVisible();
+    await expect(page).toHaveURL(/\/trip\/okinawa-trip-2026-Ray\/add-stop\?day=\d+/);
+    await expect(page.getByTestId('add-stop-page')).toBeVisible();
   });
 
-  test('modal 含 3 tab + 5 category subtab', async ({ page }) => {
-    await page.goto('/trip/okinawa-trip-2026-Ray');
-    await page.getByTestId('trip-add-stop-trigger').click();
+  test('page 含 3 tab + 5 category subtab', async ({ page }) => {
+    await page.goto('/trip/okinawa-trip-2026-Ray/add-stop?day=1');
+    await expect(page.getByTestId('add-stop-page')).toBeVisible();
     // 3 tabs
     await expect(page.getByTestId('add-stop-tab-search')).toBeVisible();
     await expect(page.getByTestId('add-stop-tab-saved')).toBeVisible();
@@ -40,22 +43,18 @@ test.describe('AddStopModal — Section 3', () => {
   });
 
   test('切到收藏 tab → render 收藏 grid 或 empty state', async ({ page }) => {
-    await page.goto('/trip/okinawa-trip-2026-Ray');
-    await page.getByTestId('trip-add-stop-trigger').click();
+    await page.goto('/trip/okinawa-trip-2026-Ray/add-stop?day=1');
     await page.getByTestId('add-stop-tab-saved').click();
-    // mock initialSavedPois 預設 [] → mockup empty state「還沒收藏景點」
     await expect(page.getByText(/還沒收藏景點/)).toBeVisible();
   });
 
   test('自訂 tab → form fields render + counter 隨 title 更新', async ({ page }) => {
-    await page.goto('/trip/okinawa-trip-2026-Ray');
-    await page.getByTestId('trip-add-stop-trigger').click();
+    await page.goto('/trip/okinawa-trip-2026-Ray/add-stop?day=1');
     await page.getByTestId('add-stop-tab-custom').click();
     await expect(page.getByTestId('add-stop-custom-title')).toBeVisible();
     await expect(page.getByTestId('add-stop-custom-time')).toBeVisible();
     await expect(page.getByTestId('add-stop-custom-duration')).toBeVisible();
     await expect(page.getByTestId('add-stop-custom-note')).toBeVisible();
-    // mockup-parity-qa-fixes: counter 0 也顯示「已選 0 個 · 將加入 ...」（mockup section 14:6518）
     await expect(page.getByTestId('add-stop-counter')).toContainText('已選');
     await expect(page.getByTestId('add-stop-counter')).toContainText('將加入');
     await page.getByTestId('add-stop-custom-title').fill('海邊散步');
@@ -63,35 +62,41 @@ test.describe('AddStopModal — Section 3', () => {
   });
 
   test('自訂 tab 缺 title 點完成 → inline error', async ({ page }) => {
-    await page.goto('/trip/okinawa-trip-2026-Ray');
-    await page.getByTestId('trip-add-stop-trigger').click();
+    await page.goto('/trip/okinawa-trip-2026-Ray/add-stop?day=1');
     await page.getByTestId('add-stop-tab-custom').click();
     await page.getByTestId('add-stop-confirm').click();
     await expect(page.getByTestId('add-stop-custom-error')).toBeVisible();
     await expect(page.getByTestId('add-stop-custom-error')).toContainText('請輸入');
   });
 
-  test('Esc key → modal 關閉', async ({ page }) => {
+  test('TitleBar 返回按鈕 → 回到 trip 頁面', async ({ page }) => {
     await page.goto('/trip/okinawa-trip-2026-Ray');
     await page.getByTestId('trip-add-stop-trigger').click();
-    await expect(page.getByTestId('add-stop-modal')).toBeVisible();
-    await page.keyboard.press('Escape');
-    await expect(page.getByTestId('add-stop-modal')).not.toBeVisible();
-  });
-
-  test('close button → modal 關閉', async ({ page }) => {
-    await page.goto('/trip/okinawa-trip-2026-Ray');
-    await page.getByTestId('trip-add-stop-trigger').click();
-    await page.getByTestId('add-stop-modal-close').click();
-    await expect(page.getByTestId('add-stop-modal')).not.toBeVisible();
+    await expect(page.getByTestId('add-stop-page')).toBeVisible();
+    // TitleBar 返回 button — TitleBar 標準 testid: titlebar-back
+    await page.getByTestId('titlebar-back').click();
+    await expect(page).toHaveURL(/\/trip\/okinawa-trip-2026-Ray$|\/trips/);
   });
 
   test('subtab 切 food → .is-active 套到該 subtab', async ({ page }) => {
-    await page.goto('/trip/okinawa-trip-2026-Ray');
-    await page.getByTestId('trip-add-stop-trigger').click();
+    await page.goto('/trip/okinawa-trip-2026-Ray/add-stop?day=1');
     await expect(page.getByTestId('add-stop-subtab-all')).toHaveClass(/is-active/);
     await page.getByTestId('add-stop-subtab-food').click();
     await expect(page.getByTestId('add-stop-subtab-food')).toHaveClass(/is-active/);
     await expect(page.getByTestId('add-stop-subtab-all')).not.toHaveClass(/is-active/);
+  });
+
+  test('TitleBar 完成按鈕 (responsive icon+text/icon-only) 與 bottom bar 同步 disabled', async ({ page }) => {
+    await page.goto('/trip/okinawa-trip-2026-Ray/add-stop?day=1');
+    // search tab default 0 selected → 兩處 confirm 都 disabled
+    const titleBarConfirm = page.getByTestId('add-stop-titlebar-confirm');
+    const bottomConfirm = page.getByTestId('add-stop-confirm');
+    await expect(titleBarConfirm).toBeDisabled();
+    await expect(bottomConfirm).toBeDisabled();
+    // switch to custom + fill title → enable
+    await page.getByTestId('add-stop-tab-custom').click();
+    await page.getByTestId('add-stop-custom-title').fill('測試景點');
+    await expect(titleBarConfirm).toBeEnabled();
+    await expect(bottomConfirm).toBeEnabled();
   });
 });
